Remove the bridge and use only a thing / thing handler
