--- bundles/org.openhab.binding.solax/src/main/java/org/openhab/binding/solax/internal/InverterDataUpdateListener.java
@@ -1,26 +1,0 @@
-/**
- * Copyright (c) 2010-2023 Contributors to the openHAB project
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0
- *
- * SPDX-License-Identifier: EPL-2.0
- */
-package org.openhab.binding.solax.internal;
-
-import org.eclipse.jdt.annotation.NonNullByDefault;
-import org.openhab.binding.solax.internal.model.InverterData;
-
-/**
- * The {@link InverterDataUpdateListener} is used to populate updated data to the interested parties
- *
- * @author Konstantin Polihronov - Initial contribution
- */
-@NonNullByDefault
-public interface InverterDataUpdateListener {
-    void updateListener(InverterData data);
-}
