+++ bundles/org.openhab.binding.solax/src/main/resources/OH-INF/i18n/solax.properties
@@ -1,7 +1,30 @@
 # thing types
 
-thing-type.solax.inverter.label = Default Inverter Label
-thing-type.solax.inverter.description = The inverter representation in OpenHAB
+thing-type.solax.localConnectInverter.label = Local Connect Inverter
+thing-type.solax.localConnectInverter.description = The inverter representation that supports local connections via HTTP
+
+# thing types config
+
+thing-type.config.solax.localConnectInverter.hostname.label = Network Address
+thing-type.config.solax.localConnectInverter.hostname.description = IP address or the host name of the Wi-Fi module
+thing-type.config.solax.localConnectInverter.password.label = Password
+thing-type.config.solax.localConnectInverter.password.description = Password for accessing the Wi-Fi module (the serial number of the Wi-Fi module)
+thing-type.config.solax.localConnectInverter.refreshInterval.label = Refresh Interval
+thing-type.config.solax.localConnectInverter.refreshInterval.description = Specifies the refresh interval in seconds.
+
+# channel types
+
+channel-type.solax.battery-temperature.label = Battery Temperature
+channel-type.solax.battery-temperature.description = Battery Temperature
+channel-type.solax.frequency.label = Electric Frequency
+channel-type.solax.frequency.description = Electric frequency of the input/output voltage of the inverter
+channel-type.solax.last-retrieve-time-stamp.label = Last retrieve time stamp
+channel-type.solax.last-retrieve-time-stamp.description = Last time with a successful retrieval of data
+channel-type.solax.raw-data-type.label = Raw Data
+channel-type.solax.raw-data-type.description = Raw data retrieved from the system as a JSON
+
+# thing types
+
 thing-type.solax.localConnect.label = Solax Local Wifi bridge
 thing-type.solax.localConnect.description = Solax direct connect to the local Wifi bridge
 
@@ -14,16 +37,10 @@ thing-type.config.solax.localConnect.password.description = Password for accessi
 thing-type.config.solax.localConnect.refreshInterval.label = Refresh Interval
 thing-type.config.solax.localConnect.refreshInterval.description = Specifies the refresh interval in seconds.
 
-# channel types
+# thing types
 
-channel-type.solax.battery-temperature.label = Battery Temperature
-channel-type.solax.battery-temperature.description = Battery Temperature
-channel-type.solax.frequency.label = Electric Frequency
-channel-type.solax.frequency.description = Electric Frequency of the input/output voltage of the inverter
-channel-type.solax.last-retrieve-time-stamp.label = Last retrieve time stamp
-channel-type.solax.last-retrieve-time-stamp.description = Last time with a successful retrieval of data
-channel-type.solax.raw-data-type.label = Raw Data
-channel-type.solax.raw-data-type.description = Raw data retrieved from the system as a JSON
+thing-type.solax.inverter.label = Default Inverter Label
+thing-type.solax.inverter.description = The inverter representation in OpenHAB
 
 # thing types config
