+++ bundles/org.openhab.binding.solax/src/main/resources/OH-INF/thing/localConnectInverter.xml
@@ -4,14 +4,10 @@
 	xmlns:thing="https://openhab.org/schemas/thing-description/v1.0.0"
 	xsi:schemaLocation="https://openhab.org/schemas/thing-description/v1.0.0 https://openhab.org/schemas/thing-description-1.0.0.xsd">
 
-	<thing-type id="inverter">
+	<thing-type id="localConnectInverter">
 
-		<supported-bridge-type-refs>
-			<bridge-type-ref id="localConnect"/>
-		</supported-bridge-type-refs>
-
-		<label>Default Inverter Label</label>
-		<description>The inverter representation in OpenHAB</description>
+		<label>Local Connect Inverter</label>
+		<description>The inverter representation that supports local connections via HTTP</description>
 
 		<channels>
 			<channel id="inverter-output-power" typeId="system.electric-power"/>
@@ -50,5 +46,22 @@
 			<property name="serialNumber"/>
 		</properties>
 
+		<config-description>
+			<parameter name="refreshInterval" type="integer" min="1" max="600">
+				<label>Refresh Interval</label>
+				<description>Specifies the refresh interval in seconds.</description>
+				<default>10</default>
+			</parameter>
+			<parameter name="password" type="text" required="true">
+				<label>Password</label>
+				<description>Password for accessing the Wi-Fi module (the serial number of the Wi-Fi module)</description>
+				<context>password</context>
+			</parameter>
+			<parameter name="hostname" type="text" required="true">
+				<label>Network Address</label>
+				<description>IP address or the host name of the Wi-Fi module</description>
+				<context>network-address</context>
+			</parameter>
+		</config-description>
 	</thing-type>
 </thing:thing-descriptions>
